[#1] Common package refactoring to better reflect Clean Architecture approach.
