+++ shared/src/commonMain/kotlin/dev/epool/hellokmp/di/dependencies.kt
@@ -1,8 +1,8 @@
 package dev.epool.hellokmp.di
 
-import dev.epool.hellokmp.network.networkModule
+import dev.epool.hellokmp.repository.network.networkModule
 import dev.epool.hellokmp.repository.repositoryModule
-import dev.epool.hellokmp.usecases.useCaseModule
+import dev.epool.hellokmp.domain.useCaseModule
 import org.kodein.di.Kodein
 
 internal val myKodein = Kodein.lazy {
